@@ -1,6 +1,7 @@
 plugins {
     id 'com.android.application'
     id 'com.google.gms.google-services'
+    id 'org.jetbrains.kotlin.android'
 }
 
 android {
@@ -16,7 +17,7 @@ android {
             storePassword keystoreProperties['store.password']
         }
     }
-    compileSdk = 34
+    compileSdk 35
 
     defaultConfig {
         applicationId "com.bao.singaporeotcalculator"
@@ -54,6 +55,9 @@ android {
         unitTests.includeAndroidResources = true
     }
     namespace 'com.bao.singaporeotcalculator'
+    kotlinOptions {
+        jvmTarget = '17'
+    }
 }
 
 dependencies {
@@ -61,6 +65,7 @@ dependencies {
     implementation 'androidx.annotation:annotation:1.8.2'
     implementation 'com.google.android.material:material:1.12.0'
     implementation 'androidx.constraintlayout:constraintlayout:2.1.4'
+    implementation 'androidx.core:core-ktx:1.15.0'
     testImplementation 'junit:junit:4.13.2'
     testImplementation 'androidx.test:core:1.6.1'
     testImplementation 'org.mockito:mockito-core:1.10.19'
